+++ packages/fighting-design/space/src/space.vue
@@ -1,39 +1,24 @@
 <script lang="ts" setup name="FSpace">
   import { Props } from './props'
-  import { computed } from 'vue'
-  import type { CSSProperties } from 'vue'
-  import type { ClassList } from '../../_interface'
+  import { useList } from '../../_hooks'
 
   const prop = defineProps(Props)
+
+  const { classes, styles } = useList(prop, 'space')
 
   /**
    * 样式列表
    */
-  const classList = computed((): ClassList => {
-    const { wrap, vertical, spacing } = prop
-
-    return [
-      'f-space',
-      {
-        [`f-space__${spacing}`]: spacing,
-        'f-space__wrap': wrap,
-        'f-space__vertical': vertical
-      }
-    ] as const
-  })
+  const classList = classes(['spacing', 'nowrap', 'vertical'], 'f-space')
 
   /**
    * 类名列表
    */
-  const styleList = computed((): CSSProperties => {
-    const { rowGap, columnGap } = prop
-
-    return { rowGap, columnGap } as const
-  })
+  const styleList = styles(['rowGap', 'columnGap'])
 </script>
 
 <template>
-  <div v-if="$slots.default" :class="classList" :style="styleList">
+  <div v-if="$slots.default" role="none" :class="classList" :style="styleList">
     <slot />
   </div>
 </template>