+++ start/src/App.vue
@@ -1,10 +1,5 @@
-<script lang="ts" setup>
-  import { FIconApps, FIconBlock } from '@fighting-design/fighting-icon'
-</script>
+<script lang="ts" setup></script>
 
-<template>
-  <f-space>
-    <f-button circle :before-icon="FIconApps" type="primary" />
-    <!-- <f-button type="success" :before-icon="FIconApps" :after-icon="FIconBlock"> 前后都有 </f-button> -->
-  </f-space>
-</template>
+<template></template>
+
+<style lang="scss" scoped></style>