+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+**新增**
+
 - 新增 `f-button` 组件 `spread` 配置项
+
+**修复**
+
+- 修复 `f-image-preview` 的警告错误
 
 ## 0.11.1 (2022-12-13)